pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bluespy-mcp"
version = "1.2.1"
description = "MCP server for blueSPY Bluetooth LE protocol analyzer"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Mohammad Afaneh", email = "mafaneh@novelbits.io" },
]
keywords = ["bluetooth", "ble", "mcp", "bluespy", "sniffer", "protocol-analyzer"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "fastmcp>=2.0.0",
]

[project.scripts]
bluespy-mcp = "bluespy_mcp.server:main"

[project.urls]
Homepage = "https://github.com/novelbits/bluespy-mcp"
Repository = "https://github.com/novelbits/bluespy-mcp"
Issues = "https://github.com/novelbits/bluespy-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/bluespy_mcp"]

[project.optional-dependencies]
e2e = ["claude-code-sdk", "pytest-asyncio"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-m 'not e2e'"
asyncio_mode = "auto"
markers = [
    "hardware: tests requiring blueSPY hardware and dylib (deselect with '-m \"not hardware\"')",
    "e2e: end-to-end tests using Claude agent SDK (deselect with '-m \"not e2e\"')",
    "file_only: e2e tests that only need capture files, no hardware",
    "no_hardware: tests that require NO hardware connected (negative path)",
]
